+++ src/HamstasKitties.Shared/Core/Interfaces/IAccelerometerService.cs
@@ -1,0 +1,16 @@
+using Microsoft.Xna.Framework;
+
+namespace HamstasKitties.Core.Interfaces;
+
+public interface IAccelerometerService : IManager
+{
+    bool IsReading { get; }
+    Vector3 CurrentAcceleration { get; }
+
+    event EventHandler<AccelerometerEventArgs>? OnShakeDetected;
+}
+
+public class AccelerometerEventArgs : EventArgs
+{
+    public Vector3 Acceleration { get; set; }
+}