+++ src/HamstasKitties.Shared/Core/Interfaces/IManager.cs
@@ -1,0 +1,7 @@
+namespace HamstasKitties.Core.Interfaces;
+
+public interface IManager
+{
+    bool Initialize();
+    bool Finalize();
+}