+++ src/HamstasKitties.Shared/Core/Interfaces/IVibratorService.cs
@@ -1,0 +1,14 @@
+namespace HamstasKitties.Core.Interfaces;
+
+public interface IVibratorService : IManager
+{
+    bool IsEnabled { get; set; }
+    void Vibrate(VibrationDuration duration);
+}
+
+public enum VibrationDuration
+{
+    Short = 100,
+    Medium = 500,
+    Long = 1000
+}